+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 
-## Changelog
+# Changelog
 
-### 1.0.0 - 2017-02-18
+## 1.1.0 - 2017-02-21
+
+- Support Multiple GeoJSON Polygon
+- Added GeoJSON input support
+
+## 1.0.0 - 2017-02-18
 
 - Stable release
 - Initialize library